+++ src/app/modules/search-parameters/search-parameters.component.ts
@@ -1,20 +1,31 @@
-import { Component, ViewChildren, QueryList } from '@angular/core';
-import { AbstractControl, FormArray, FormControl } from '@angular/forms';
+import { Component, ViewChildren, QueryList, ViewChild } from '@angular/core';
+import { FormControl } from '@angular/forms';
 import {
   BaseControlValueAccessor,
   createControlValueAccessorProviders
 } from '../base-control-value-accessor';
 import { SearchParameter } from 'src/app/types/search.parameter';
-import { SearchCondition } from '../../types/search.condition';
-import { SearchParameterGroupComponent } from '../search-parameter-group/search-parameter-group.component';
 import { ErrorStateMatcher } from '@angular/material/core';
 import { ErrorManager } from '../../shared/error-manager/error-manager.service';
+import { FhirBackendService } from '../../shared/fhir-backend/fhir-backend.service';
+import { map, take } from 'rxjs/operators';
 import {
-  ConnectionStatus,
-  FhirBackendService
-} from '../../shared/fhir-backend/fhir-backend.service';
-import { filter } from 'rxjs/operators';
+  Field,
+  Option,
+  QueryBuilderComponent,
+  QueryBuilderConfig,
+  Rule,
+  RuleSet
+} from '../../../query-builder/public-api';
+import { Observable } from 'rxjs';
+import {
+  AutocompleteComponent,
+  AutocompleteOption
+} from '../autocomplete/autocomplete.component';
 import { LiveAnnouncer } from '@angular/cdk/a11y';
+import { SearchParameterComponent } from '../search-parameter/search-parameter.component';
+import { MatButton } from '@angular/material/button';
+import { ResourceTypeCriteria } from '../../types/search-parameters';
 
 /**
  * Component for managing resources search parameters
@@ -34,50 +45,134 @@ import { LiveAnnouncer } from '@angular/cdk/a11y';
 export class SearchParametersComponent extends BaseControlValueAccessor<
   SearchParameter[]
 > {
-  @ViewChildren(SearchParameterGroupComponent)
-  searchParameterGroupComponents: QueryList<SearchParameterGroupComponent>;
-  parameterGroupList = new FormArray([]);
+  @ViewChildren(AutocompleteComponent)
+  resourceTypeComponents: QueryList<AutocompleteComponent>;
+  @ViewChildren(SearchParameterComponent)
+  searchParameterComponents: QueryList<SearchParameterComponent>;
+  @ViewChildren('addCriterionBtn')
+  buttons: QueryList<MatButton>;
+  @ViewChild(QueryBuilderComponent)
+  queryBuilderComponent: QueryBuilderComponent;
+  public queryCtrl: FormControl = new FormControl({});
+  public queryBuilderConfig: QueryBuilderConfig = { fields: {} };
+  resourceTypes$: Observable<AutocompleteOption[]>;
+  selectedSearchParameterNamesMap = new Map<ResourceTypeCriteria, string[]>();
 
   constructor(
     private fhirBackend: FhirBackendService,
     private liveAnnoncer: LiveAnnouncer
   ) {
     super();
-    this.parameterGroupList.valueChanges.subscribe((value) => {
-      this.onChange(value);
-    });
 
-    fhirBackend.initialized
-      .pipe(filter((status) => status === ConnectionStatus.Ready))
-      .subscribe(() => {
-        // Clear search parameters on server change
-        this.parameterGroupList.clear();
-      });
-  }
-
-  /**
-   * Add new search parameter group to search parameter group list
-   */
-  public addParameterGroup(): void {
-    this.parameterGroupList.push(
-      new FormControl({
-        resourceType: '',
-        parameters: []
-      })
+    this.resourceTypes$ = fhirBackend.currentDefinitions$.pipe(
+      map((definitions) => Object.keys(definitions.resources))
     );
-    this.liveAnnoncer.announce('A new line of resource type is added.');
-    // Focus the input control of the newly added resource type line.
-    setTimeout(() => {
-      this.searchParameterGroupComponents.last.focusResourceTypeInput();
-    }, 0);
+
+    fhirBackend.currentDefinitions$.subscribe((definitions) => {
+      // Clear search parameters on server change
+      const config = {
+        allowEmptyRulesets: true,
+        fields: {},
+        /**
+         * Adds a rule (criterion) for a resource type
+         */
+        addRule: (parent: RuleSet): void => {
+          parent.rules = parent.rules.concat([
+            {
+              field: {}
+            } as Rule
+          ]);
+          this.liveAnnoncer.announce(
+            'A new line of search criterion is added.'
+          );
+          // Focus the input control of the newly added search parameter line.
+          this.searchParameterComponents.changes
+            .pipe(take(1))
+            .subscribe((components) => {
+              setTimeout(() => components.last.focusSearchParamNameInput());
+            });
+        },
+        /**
+         * Removes a rule (criterion) from a resource type criteria
+         * @param rule - criterion
+         * @param parent - resource type criteria
+         */
+        removeRule: (rule: Rule, parent: RuleSet) => {
+          parent.rules = parent.rules.filter((r) => r !== rule);
+          if ('resourceType' in parent) {
+            this.updateSelectedSearchParameterNames(
+              (parent as unknown) as ResourceTypeCriteria
+            );
+          } else if ('resourceType' in rule) {
+            this.selectedSearchParameterNamesMap.delete(
+              (rule as unknown) as ResourceTypeCriteria
+            );
+          }
+        },
+        getInputType: (fieldName: string, operator: string): string => {
+          return 'search-parameter';
+        },
+        getOperators: (fieldName: string, field: Field): string[] => {
+          return [];
+        },
+        // Override to an empty method only to remove the exception
+        getOptions: (fieldName: string): Option[] => null
+      };
+      const resourceTypes = Object.keys(definitions.resources);
+      resourceTypes.forEach((resourceType) => {
+        definitions.resources[resourceType].searchParameters.forEach((desc) => {
+          config.fields[resourceType + '-' + desc.element] = desc;
+        });
+      });
+      this.queryCtrl.setValue({
+        condition: 'and',
+        rules: []
+      });
+      this.queryBuilderConfig = config;
+    });
   }
 
   /**
-   * Remove search parameter group from search parameter group list
+   * Focus the input control of the newly added add criterion button.
    */
-  public removeParameterGroup(item: AbstractControl): void {
-    this.parameterGroupList.removeAt(
-      this.parameterGroupList.controls.indexOf(item)
+  focusOnAddCriterionBtn(): void {
+    const prevButtons = this.buttons.toArray();
+    this.buttons.changes.pipe(take(1)).subscribe((components) => {
+      setTimeout(() =>
+        components.find((btn) => prevButtons.indexOf(btn) === -1).focus()
+      );
+    });
+  }
+
+  /**
+   * Adds a ruleset for a resource type
+   * @param ruleset parent ruleset
+   */
+  addResourceType(ruleset: RuleSet): void {
+    const newResourceTypeCriteria = {
+      condition: 'and',
+      rules: [],
+      // RuleSet is treated as a ruleset for a resource type
+      // if it has a "resourceType" property
+      resourceType: ''
+    };
+
+    ruleset.rules = ruleset.rules.concat(newResourceTypeCriteria as RuleSet);
+
+    this.liveAnnoncer.announce(
+      'A new field for selecting a record type has been added.'
+    );
+
+    // Focus the input control of the newly added resource type line.
+    this.resourceTypeComponents.changes
+      .pipe(take(1))
+      .subscribe((components) => {
+        setTimeout(() => components.last.focus());
+      });
+
+    this.selectedSearchParameterNamesMap.set(
+      newResourceTypeCriteria as ResourceTypeCriteria,
+      []
     );
   }
 
@@ -85,19 +180,25 @@ export class SearchParametersComponent extends BaseControlValueAccessor<
     // TODO
   }
 
-  // Get search conditions from each row
-  getConditions(): SearchCondition[] {
-    const conditions = this.searchParameterGroupComponents
-      .map((c) => c.getConditions())
-      // Filter out empty resource type or criteria
-      .filter((c) => c.resourceType && c.criteria);
-    if (!conditions.some((c) => c.resourceType === 'Patient')) {
-      // add default Patient condition if missing
-      conditions.push({
-        resourceType: 'Patient',
-        criteria: ''
-      });
-    }
-    return conditions;
+  /**
+   * Returns the indefinite article ('a' or 'an') for the specified word.
+   */
+  getIndefiniteArticle(word: string): string {
+    return /^[euioa]/i.test(word) ? 'an' : 'a';
+  }
+
+  /**
+   * Updates the list of already selected FHIR search parameter names for the
+   * specified resource type criteria. This list is used to exclude dropdown
+   * options to avoid duplicate criteria.
+   * @param parentRuleSet - resource type criteria
+   */
+  updateSelectedSearchParameterNames(
+    parentRuleSet: ResourceTypeCriteria
+  ): void {
+    this.selectedSearchParameterNamesMap.set(
+      parentRuleSet,
+      parentRuleSet.rules.map((c) => c.field.element)
+    );
   }
 }