+++ src/app/shared/fhir-backend/fhir-backend.service.ts
@@ -17,6 +17,7 @@ import { FhirServerFeatures } from '../../types/fhir-server-features';
 import { escapeStringForRegExp } from '../utils';
 import { SettingsService } from '../settings-service/settings.service';
 import { find } from 'lodash-es';
+import { filter, map } from 'rxjs/operators';
 
 // RegExp to modify the URL of requests to the FHIR server.
 // If the URL starts with the substring "$fhir", it will be replaced
@@ -124,9 +125,14 @@ export class FhirBackendService implements HttpBackend {
     this.fhirClient = new FhirBatchQuery({
       serviceBaseUrl: queryServer || defaultServer
     });
+    this.currentDefinitions$ = this.initialized.pipe(
+      filter((status) => status === ConnectionStatus.Ready),
+      map(() => this.getCurrentDefinitions())
+    );
   }
   // Whether the connection to server is initialized.
   initialized = new BehaviorSubject(ConnectionStatus.Pending);
+  currentDefinitions$: Observable<any>;
 
   // Whether to cache requests to the FHIR server
   private isCacheEnabled = true;