+++ src/query-builder/lib/query-builder/query-switch-group.directive.ts
@@ -1,0 +1,10 @@
+/**
+ * Copied from https://github.com/zebzhao/Angular-QueryBuilder with modifications.
+ * See ../../query-builder-license.md
+ */
+import { Directive, TemplateRef } from '@angular/core';
+
+@Directive({selector: '[querySwitchGroup]'})
+export class QuerySwitchGroupDirective {
+  constructor(public template: TemplateRef<any>) {}
+}