+++ src/query-builder/public-api.ts
@@ -1,0 +1,25 @@
+/**
+ * Copied from https://github.com/zebzhao/Angular-QueryBuilder with modifications.
+ * See ./query-builder-license.md
+ */
+/*
+ * Public API Surface of angular2-query-builder
+ */
+
+export * from './lib/query-builder/query-builder.interfaces';
+
+export * from './lib/query-builder/query-builder.component';
+
+export * from './lib/query-builder/query-button-group.directive';
+export * from './lib/query-builder/query-entity.directive';
+export * from './lib/query-builder/query-field.directive';
+export * from './lib/query-builder/query-input.directive';
+export * from './lib/query-builder/query-operator.directive';
+export * from './lib/query-builder/query-switch-group.directive';
+export * from './lib/query-builder/query-switch-group-prefix.directive';
+export * from './lib/query-builder/query-tree-container-suffix.directive';
+export * from './lib/query-builder/query-remove-button.directive';
+export * from './lib/query-builder/query-empty-warning.directive';
+export * from './lib/query-builder/query-arrow-icon.directive';
+
+export * from './lib/angular2-query-builder.module';